requirements.txt
requests-cache
natsort
tqdm
pillow
python-dotenv